pyproject.toml
[project]
name = "cdp-python-backend-template"
version = "0.1.0"
description = "CDP Python Backend Template"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asgi-logger==0.1.0",
    "aiobotocore>=2.13.0",
    "asyncpg>=0.30.0",
    "aws-embedded-metrics==3.3.0",
    "dnspython==2.7.0",
    "ecs-logging==2.2.0",
    "fastapi==0.129.0",
    "httpx==0.28.1",
    "pydantic-settings==2.9.1",
    "python-multipart>=0.0.12",
    "pyjwt>=2.10.0",
    "uvicorn==0.34.3",
]

[dependency-groups]
dev = [
    "pre-commit==4.2.0",
    "pytest==8.4.0",
    "pytest-asyncio==1.0.0",
    "pytest-cov==6.2.1",
    "pytest-mock==3.15.1",
    "ruff==0.11.13",
    "taskipy==1.14.1",
]

[tool.taskipy.tasks]
format = "uv run ruff format . && uv run ruff check . --fix"
